requirements.txt
torch>=1.12
torchvision>=0.13

opencv-python>=4.5
albumentations>=1.3.0

pandas>=1.3
scikit-learn>=1.0

matplotlib>=3.4
seaborn>=0.11
tqdm>=4.64

pytesseract>=0.3

imageio
Pillow>=8.2

requests>=2.25
ultralytics>=8.0.0
gitpython>=3.1
thop>=0.1.1
